issue #16 Feat: (data) duplicate_email_check from server
